Refactor device settings into smaller components
