+++ src/routes/DeviceSettingsPage/containers/MorpheusSettingsForm.js
@@ -1,7 +1,6 @@
 import { connect } from 'react-redux'
 import { compose } from 'redux'
 import { reduxForm } from 'redux-form'
-import { push } from 'react-router-redux'
 import validator from '@helpers/validator'
 import schema from '@schemas/addMorpheus'
 import MorpheusSettingsForm from '@routes/DeviceSettingsPage/components/MorpheusSettingsForm'
@@ -10,19 +9,8 @@ import * as morpheusActions from '@modules/morpheus/actions'
 const validate = values => validator(values, schema)
 
 const mapDispatchToProps = dispatch => ({
-  deleteMorpheus(morpheus) {
-    dispatch(morpheusActions.deleteMorpheus(morpheus))
-      .then((success) => {
-        if (success) {
-          dispatch(push('/device-settings/morpheus'))
-        }
-      })
-  },
   updateMorpheus(morpheus) {
     dispatch(morpheusActions.updateMorpheus(morpheus))
-  },
-  clearError() {
-    dispatch(morpheusActions.clearMorpheusErrors())
   },
 })
 
@@ -31,8 +19,6 @@ const mapStateToProps = (state, ownProps) => ({
     resend: ownProps.serial ? state.morpheus.get('morpheus').get(ownProps.serial).resend : null,
   },
   morpheus: ownProps.serial ? state.morpheus.get('morpheus').get(ownProps.serial) : null,
-  morpheusError: state.morpheus.get('error'),
-  morpheusRemoving: state.morpheus.get('isRemoving'),
   morpheusUpdating: state.morpheus.get('isUpdating'),
 })
